Revert PR # 4543 (api client changes) (#4636)
* Revert "Split out Tamanu API client (#4543)"

This reverts commit f365e45afa7d1faf5c00ef76d990085f4d24dc8d.

* Revert "TAMOC-54: Fix broken builds due to @tamanu/api-client not building (#4626)"

This reverts commit 359571798f78797889d4a65ee73d3494933b0f78.
