+++ packages/.new-package/tsconfig.json
@@ -5,5 +5,6 @@
   ],
   "compilerOptions": {
     "baseUrl": "./src",
+    "tsBuildInfoFile": "./dist/.tsbuildinfo",
   }
 }
