--- packages/api-client/.gitignore
@@ -1,2 +1,0 @@
-dist
-*.tsbuildinfo
