--- packages/api-client/.swcrc
@@ -1,16 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/swcrc",
-  "exclude": "node_modules/",
-  "module": {
-    "type": "commonjs"
-  },
-  "jsc": {
-    "parser": {
-      "syntax": "typescript",
-      "tsx": true,
-      "dts": true
-    },
-    "target": "es2020",
-    "keepClassNames": true
-  }
-}
