--- packages/api-client/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@tamanu/api-client",
-  "version": "1.30.0",
-  "private": true,
-  "description": "API client for Tamanu Facility Server",
-  "exports": {
-    ".": "./dist/index.js"
-  },
-  "types": "./dist/index.d.ts",
-  "homepage": "https://github.com/beyondessential/tamanu.git#readme",
-  "repository": "git@github.com:beyondessential/tamanu.git",
-  "author": "Beyond Essential Systems Pty. Ltd.",
-  "license": "SEE LICENSE IN ../../license",
-  "scripts": {
-    "build": "yarn run build:src && yarn run build:types",
-    "build:src": "swc --delete-dir-on-start --out-dir dist --copy-files --source-maps true src",
-    "build:types": "tsc --declaration --emitDeclarationOnly --noEmit false && move-dts src dist",
-    "build-watch": "yarn run build && yarn run build:src --watch",
-    "test": "echo no tests yet",
-    "lint": "tsc",
-    "start": "node dist/cli.js"
-  },
-  "devDependencies": {
-    "@tamanu/build-tooling": "*",
-    "@types/node": "^18.14.6"
-  },
-  "dependencies": {
-    "@casl/ability": "^4.1.0",
-    "@tamanu/constants": "*",
-    "@tamanu/shared": "*",
-    "qs": "^6.10.2"
-  }
-}
