--- packages/api-client/src/index.ts
@@ -1,23 +1,0 @@
-export {
-  AuthExpiredError,
-  ServerResponseError,
-  ServerUnavailableError,
-  VersionIncompatibleError,
-} from './errors';
-export type {
-  FetchImplementation,
-  ResponseError,
-  RequestOptions,
-} from './fetch';
-export { setFetchImplementation } from './fetch';
-
-export type {
-  UserResponse,
-  AuthFailureHandler,
-  VersionIncompatibleHandler,
-  QueryData,
-  FetchConfig,
-  ChangePasswordArgs,
-  LoginOutput,
-} from './TamanuApi';
-export { TamanuApi } from './TamanuApi';
