--- packages/api-client/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": ["../../common.tsconfig.json"],
-  "include": [
-    "src/**/*.ts",
-    "package.json"
-  ],
-  "compilerOptions": {
-    "baseUrl": "./src",
-    "resolveJsonModule": true
-  }
-}
